+++ assets/CHANGELOG.md
@@ -1,3 +1,8 @@
+## v0.2.0 - 2025-11-30 UTC
+
+- Added `Exclusion List` configuration to specify item names that should not be collected from the cruiser.
+- Removed `Easter egg` from the built-in and default exclusion list. It was bad for QoL rather than spoiling.
+
 ## v0.1.0 - 2025-11-27 UTC
 
 - Initial release on Thunderstore.